requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-embed-video==1.4.10
idna==3.11
pillow==12.0.0
razorpay==2.0.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
